example/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiovban-examples"
version = "0.1.0"
authors = [
  { name="William Best", email="wmbest2@gmail.com" },
]
description = "An asyncio VBAN implementation"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "aiovban",
    "aiovban_pyaudio"
]


[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
aiovban-receiver = "src:audio_receiver.py"